Move user profile under current user widget
