+++ web/src/app/shared/components/header-layout/current-user-widget/user-profile-popup/user-profile-popup.component.html
