+++ web/src/app/shared/components/header-layout/current-user-widget/user-profile-popup/user-profile-popup.module.ts
@@ -18,7 +18,7 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { MatButtonModule } from '@angular/material/button';
 import { MatDialogModule } from '@angular/material/dialog';
-import { UserProfilePopupComponent } from 'app/components/user-profile-popup/user-profile-popup.component';
+import { UserProfilePopupComponent } from 'app/shared/components/header-layout/current-user-widget/user-profile-popup/user-profile-popup.component';
 import { UserAvatarModule } from 'app/components/user-avatar/user-avatar.module';
 
 @NgModule({
